Added controller service to grpc server
