+++ Makefile
@@ -12,4 +12,4 @@ generate-python-code: ## Generate python code from .proto
 	@echo "generating shared code";
 	@python3 -m grpc_tools.protoc -I protos --python_out=.  --pyi_out=. --grpc_python_out=. ./protos/kubekarma/shared/pb2/*.proto;
 	@echo "generating controller code" ;
-	@python3 -m grpc_tools.protoc -I protos --python_out=.  --pyi_out=. --grpc_python_out=. ./protos/kubekarma/controlleroperator/grpc/pb2/*.proto
+	@python3 -m grpc_tools.protoc -I protos --python_out=.  --pyi_out=. --grpc_python_out=. ./protos/kubekarma/controlleroperator/grpcsrv/pb2/*.proto
