+++ chart/templates/service.yaml
@@ -12,5 +12,5 @@ spec:
   ports:
       - protocol: TCP
         port: {{ .Values.controller.grpc.service.port }}
-        targetPort: grpc
-        name: grpc
+        targetPort: grpcsrv
+        name: grpcsrv
