+++ chart/values.yaml
@@ -21,5 +21,5 @@ controller:
     address: "[::]"
     port: 8080
     service:
-      # @controller.grpc.service.port the exposed port by the service
+      # @controller.grpcsrv.service.port the exposed port by the service
       port: 8080
